+++ index.js
@@ -13,11 +13,10 @@ import { kitAssemblyScreen } from "./src/pages/homeCollection/kitAssembly.js";
 import { printLabelsScreen } from "./src/pages/homeCollection/printLabels.js";
 import { assignKitsScreen } from "./src/pages/homeCollection/assignKits.js";
 import { kitsReceiptScreen } from "./src/pages/homeCollection/kitsReceipt.js";
-// import { displayKitStatusReportsScreen } from "./src/pages/homeCollection/kitStatusReports.js"; // TODO: This will be added back in once the new kitStatusReports page is created
+import { displayKitStatusReportsScreen } from "./src/pages/homeCollection/kitStatusReports.js";
 import { allParticipantsScreen } from "./src/pages/homeCollection/allParticipants.js";
 import { addressesPrintedScreen } from "./src/pages/homeCollection/assignKit.js";
 import { assignedScreen } from "./src/pages/homeCollection/assigned.js";
-import { displayKitStatusReportsShippedScreen } from "./src/pages/homeCollection/kitStatusReportsShipped.js";
 import { receivedKitsScreen } from "./src/pages/homeCollection/receivedKits.js";
 import { kitCsvScreen } from "./src/pages/homeCollection/kitCSV.js";
 import { kitShipmentScreen } from "./src/pages/homeCollection/kitShipment.js";
@@ -126,7 +125,9 @@ const manageRoutes = async () => {
         else if (route === "#assignKits") assignKitsScreen(auth, route);
         else if (route === "#kitsReceipt") kitsReceiptScreen(auth, route);
         else if (route === "#kitsCsv") kitCsvScreen(auth, route);
-        else if (route === "#kitStatusReports") displayKitStatusReportsShippedScreen(auth, route); // Temporarily make kitStatusReports route call displayKitStatusReportsShippedScreen
+        else if (route === "#kitStatusReports") { 
+          displayKitStatusReportsScreen(auth, route);
+        }
         else if (route === "#allParticipants") allParticipantsScreen(auth, route);
         else if (route === "#addressPrinted") addressesPrintedScreen(auth, route);
         else if (route === "#assigned") assignedScreen(auth, route);
